+++ projects/ngx-chessground-example/src/app/privacy-policy/privacy-policy.component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-privacy-policy',
-  standalone: true,
-  templateUrl: './privacy-policy.component.html',
-  styleUrls: ['./privacy-policy.component.scss'],
+	selector: 'app-privacy-policy',
+	standalone: true,
+	templateUrl: './privacy-policy.component.html',
+	styleUrls: ['./privacy-policy.component.scss'],
 })
 export class PrivacyPolicyComponent {}